Finalizzato Room + LineGraph + MainActivity * testare che effettivamente rimangano solo 5 minuti di registrazione dati nel Room DB * testare corretta indicizzazione asse x * cominciare a registrare dati solo dopo completa visualizzazione preview * rendere più leggibili scritte in RGB Values * creare string per italiano
# migliorie
* fotocamera, si può migliorare,
   - controllare `.let` se hai implementato correttamente come su Android Developers
* correggere status bar (modalità scura/chiara-land/port)
* controlla progetto github per ulteriori migliorie

+++ app/src/main/java/com/example/hueharvester/ColorData.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "color_data")
 data class ColorData(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val timestamp: Float,
+    val timestamp: Long,
     val red: Int,
     val green: Int,
     val blue: Int
